devenv: Add opencode and rsync
I had been trying hard to use goose because I like Rust
and some of the core code is good, but opencode seems to be
nicer and well maintained.

Add infra to inject bits from npm and use it to install.

rsync is needed for my dotfiles install and is generally useful.

Signed-off-by: Colin Walters <walters@verbum.org>